Commit of the 26/12/2023

+++ src/entities/Entity.java
@@ -22,7 +22,7 @@ public abstract class Entity {
 
     }
 
-    protected void initHitBox(float x,float y,float width,float height) {
+    protected void initHitBox(float x,float y,int width,int height) {
         hitBox=new Rectangle2D.Float(x,y,width,height);
     }
     /*
